academy-watch-backend/requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
azure-identity==1.21.0
azure-keyvault-secrets==4.9.0
bleach==6.2.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Deprecated==1.2.18
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.0
Flask-Limiter==3.9.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
fonttools==4.61.1
greenlet==3.2.3
griffe==1.10.0
groq==0.32.0
gunicorn==22.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
limits==5.5.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mcp==1.12.3
mdurl==0.1.2
numpy==2.4.0
pandas==2.2.3
openai==1.109.1
openai-agents==0.2.4
ordered-set==4.1.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
psycopg==3.2.9
psycopg-binary==3.2.9
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.3.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.36.2
requests==2.32.4
RestrictedPython==7.4
rich==13.9.4
rpds-py==0.26.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==3.0.2
starlette==0.47.2
stripe==14.0.1
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.3
